refactor: project structure refactor
Change organization of the files into: cmd -> actual app with the main, pkg -> all the packages used, test -> stuff for testing like the mocks.
