+++ .gitignore
@@ -31,8 +31,7 @@
 # Dependency directories (remove the comment below to include it)
 # vendor/
 
-fl
-
+cmd/fl/fl
 .task
 vendor/
 bin/
